+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Clearlog and Mail module by @AppoPro
 - Socat vector in backdoor tcp module
+- Legacycookie generator
 
 ## [v3.2.0] - 2015-06-29
 ### Fixed